Make MenuItemListFragments observe meal selection changes
I'm not sure if this is really the best design pattern, but I'm thinking it's more efficient than recreating all the fragments in the ViewPager every time a meal is selected. This also incorporates the "Not Serving" message into the list fragment which should help with the bug where the ViewPager is not always refreshed properly when the app opens.